Some cleanups

+++ include/screen_print/026000.h
@@ -12,7 +12,7 @@ void func_8003E208(Gfx **gfxP, s32 index);
 void func_8003E3F0(Gfx **gfxP, s32 arg1, s32 arg2, s32 character);
 s32 func_8003E430(struct_8008E364 *arg0, va_list args);
 s32 func_8003E4AC(void);
-s32 func_8003E4B4(struct_8008E364 *arg0, char *arg1, s32 arg2);
+s32 func_8003E4B4(struct_8008E364 *arg0, char *arg1, size_t arg2);
 s32 func_8003E69C(Gfx **gfxP, func_8003E69C_arg1 arg1, s32 arg2, s32 arg3);
 void func_8003E6D8(void);
 void func_8003E6F8(const char *fmt, ...);
